requirements.txt
pyspark==3.2.1
PyYAML==6.0
pytest==7.1.2
pandas==1.4.2